Fix format specifier warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36510 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/plus60k.c
@@ -264,7 +264,8 @@ static int set_plus60k_base(int val, void *param)
         case 0xd100:
             break;
         default:
-            log_message(plus60k_log, "Unknown PLUS60K base address $%X.", val);
+            log_message(plus60k_log, "Unknown PLUS60K base address $%X.",
+                    (unsigned int)val);
             return -1;
     }
